+++ drivers/drivers_test.go
@@ -1,4 +1,4 @@
-package duty
+package drivers
 
 import (
 	"reflect"
